update to the cloud atoi was added at the main function thefore i can't compile on the laptop

+++ parser.h
@@ -28,6 +28,6 @@ typedef struct
 } Sequence;
 
 
-size_t parseFile(char fileName[], Sequence seqArr[MAX_SEQUENCES]);
+int parseFile(char fileName[], Sequence seqArr[MAX_SEQUENCES]);
 
 #endif //UNTITLED_PARSER_H
